S-Box result back to binary,  Changed functionXOR , added endPermuation() , finished DES0, getCiphertext()
S-Box now goes from decimal result to binary result, Changed the XOR function to be more general function so that it can be used more than once and now does the XOR operation. Added the endPermuation() for the DES rounds before being XORed put through the Final Permutation and the halves being swapped. Finally added in a getCiphertext method that will return our ciphertext to be displayed in the main.java file.