added Print for main, added comparison in DES0, Created DES1.java, fixed input file
Huge Changed for this commit, added functionality for comparison to be printed out for all classes. In Main.java all you have to do is follow the format to be able to print to next sections also added in methods to write to EncryptionOutput.txt file. Input.txt changed to fit assignment specification. Completed DES0 and DES1 for encryption. TODO: Decryption for both of these and adding in Decyrption input and out put files.